Fix: sub chunk y index read error

+++ src/include/sub_chunk.h
@@ -18,7 +18,7 @@ namespace bl {
     };
 
     class sub_chunk {
-    public:
+       public:
         struct layer {
             uint8_t bits;
             uint8_t type;
@@ -33,7 +33,7 @@ namespace bl {
 
         void set_version(uint8_t version) { this->version_ = version; }
 
-        void set_y_index(uint8_t y_index) { this->y_index_ = y_index; }
+        void set_y_index(int y_index) { this->y_index_ = y_index; }
 
         void set_layers_num(uint8_t layers_num) { this->layers_num_ = layers_num; }
 
@@ -45,7 +45,7 @@ namespace bl {
 
         void push_back_layer(const layer &layer) { this->layers_.push_back(layer); }
 
-    private:
+       private:
         uint8_t version_{0xff};
         int y_index_{0};
         uint8_t layers_num_{0xff};
